Moved endOfDay function to StorageService. 2 extra fields are added to user on creation, both of type Timestamp. One tracks the last time user uploaded GameStats and the other is the next time user is allowed to upload GameStats.
